ipfs-storage/requirements.txt
base58==2.1.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==44.0.3
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-MySQL==1.6.0
idna==3.10
ipfshttpclient==0.8.0a2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multiaddr==0.0.9
mysql-connector-python==9.3.0
netaddr==1.3.0
pycparser==2.22
pycryptodome==3.22.0
PyMySQL==1.1.1
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
tenseal==0.3.16
urllib3==2.4.0
varint==1.0.2
Werkzeug==3.1.3
